requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Babel
python-dotenv
psycopg2-binary
Werkzeug
openai
gunicorn
fpdf2
flask-socketio
eventlet
